backend/requirements.txt
Django == 5.0.6
psycopg2-binary == 2.9.9
djangorestframework == 3.15.1
python-ldap == 3.4.3
django-auth-ldap == 4.8.0
djangorestframework-simplejwt == 5.3.1
drf-yasg == 1.21.7
celery == 5.4.0
hapless == 0.2.2